requirements.txt
Flask==2.0.2
PyJWT==2.8.0
Flask-JWT-Extended==4.6.0
google-generativeai==0.5.4
groq==0.8.0
python-dotenv==1.0.1
vt-py==0.18.2
requests==2.25.1